Make several instances co-installable
* Configurable LETSENCRYPT_HOST variable name
* Configurable nginx-proxy & docker-gen labels

Combined with my nginx-proxy fork HTTPS passthrough feature this allow for
easy testing configuration setup, very usefull when one has only one server
to work with.

In swarm mode we can't give names to the containers. We have to use
labels then.  And these labels have to be configurable if we want to be
able to chain several nginx-proxy instances.

Also wait for nginx-proxy and docker-gen containers instead of erroring out.
This is useful in swarm environment where the 'depends_on' compose option
isn't available.